conditional postinstall

+++ postinstall.js
@@ -22,4 +22,9 @@ switch (osType) {
 const binaryPath = path.join(binariesDir, binaryName);
 const targetPath = path.join(__dirname, 'index.node');
 
+if (!fs.existsSync(binaryPath)) {
+    console.log(`Binary not found: ${binaryPath}. Skipping setup.`);
+    process.exit(0); // Exit gracefully
+}
+
 fs.renameSync(binaryPath, targetPath);
